# Significantly optimize parsing overhead when `--since` is specified 