--- a/lib/mutant/repository.rb
+++ b/lib/mutant/repository.rb
@@ -2,19 +2,5 @@
 
 module Mutant
   module Repository
-    # Subject filter based on repository diff
-    class SubjectFilter
-      include Adamantium, Anima.new(:diff)
-
-      # Test if subject was touched in diff
-      #
-      # @param [Subject] subject
-      #
-      # @return [Boolean]
-      def call(subject)
-        diff.touches?(subject.source_path, subject.source_lines)
-      end
-
-    end # SubjectFilter
   end # Repository
 end # Mutant